+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 ### Documentation
 
+- updated readme (bba1b1d)
 - mentioning poetry in readme (2ca651d)
 
 ### New Features
 
+- moved to uv for packagin (9721249)
 - added changelog.md (be5393c)
 
 ### Other changes
 
+- enhanced script to also install current repo (ed0d826)
+- updated setup.py (6b5ae52)
+- updated project structure template (894401f)
+- exclusions to gitignore (2c39c96)
 - added /doc to pyproject.toml (60c6681)
 - fixed typo (a6b2604)
 - first commit (24a8a68)